fixup! replace SerializedIdentity type with more specific identity types. This allows a driver to support multiple owner identity types
